feat: Add functionality to merge SVN and CBS CSV data
This commit adds functionality to merge data from two different CSV files: SVN and CBS. It imports the `formatFromSvnCsv` and `formatFromCbsCsv` functions from the `helpers` module and uses them to parse the respective CSV files. Then, it merges the datasets based on the `municipality` field and stores the result in the `mergedData` variable. Finally, it logs the merged data to the console.