feat: Add writeDataToCsv helper function
This commit adds the writeDataToCsv helper function to the src/helpers/index.ts file. The function allows writing data to a CSV file. It takes an array of objects representing the data and a filename as input. The function throws an error if the data is empty, the filename is not provided, or the data is not an array. The function writes the data to the specified CSV file using Deno's writeTextFileSync function.